Roll java17 module into e2e
- Remove the separate java17 testing module in favor of e2e testing.
- Upgrade the minimum compile requirement to build the code base to
  Java 17
