+++ e2e/pom.xml
@@ -33,6 +33,9 @@
         <!-- docker tests run long -->
         <basepom.test.timeout>600</basepom.test.timeout>
         <moduleName>org.jdbi.v3.e2e</moduleName>
+        <!-- e2e tests can use all java 17 features. -->
+        <!-- current PMD 6.xx does not allow upgrade to full 21 yet -->
+        <project.build.targetJdk>17</project.build.targetJdk>
     </properties>
 
     <dependencies>
@@ -168,6 +171,16 @@
                     </plugin>
                 </plugins>
             </build>
+        </profile>
+        <!-- e2e tests require java 17 or better. Skip tests otherwise (for java 11 testing) -->
+        <profile>
+            <id>no-java17</id>
+            <activation>
+                <jdk>(,17]</jdk>
+            </activation>
+            <properties>
+                <basepom.test.skip>true</basepom.test.skip>
+            </properties>
         </profile>
     </profiles>
 </project>
