+++ e2e/src/test/java/org/jdbi/v3/e2e/java17/TestTextBlock.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.java17;
+package org.jdbi.v3.e2e.java17;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,7 +20,9 @@ import org.jdbi.v3.core.mapper.RowMappers;
 import org.jdbi.v3.core.mapper.reflect.ConstructorMapper;
 import org.jdbi.v3.sqlobject.SqlObject;
 import org.jdbi.v3.sqlobject.SqlObjectPlugin;
+import org.jdbi.v3.sqlobject.customizer.BindMethods;
 import org.jdbi.v3.sqlobject.statement.SqlQuery;
+import org.jdbi.v3.sqlobject.statement.SqlUpdate;
 import org.jdbi.v3.testing.junit5.JdbiExtension;
 import org.jdbi.v3.testing.junit5.internal.TestingInitializers;
 import org.junit.jupiter.api.BeforeEach;
@@ -29,7 +31,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class TestRecord {
+class TestTextBlock {
 
     @RegisterExtension
     JdbiExtension h2Extension = JdbiExtension.h2()
@@ -53,12 +55,7 @@ class TestRecord {
 
     @Test
     void testBinding() {
-        var handle = h2Extension.getSharedHandle();
-
-        try (var update = handle.createUpdate("INSERT INTO users (id, name) values <values>")) {
-            update.bindMethodsList("values", List.of(new User(3, "Charlie")), List.of("id", "name"))
-                    .execute();
-        }
+        dao.insertUser(new User(3, "Charlie"));
 
         var user = dao.getUser(3);
         assertThat(user).isPresent()
@@ -69,10 +66,24 @@ class TestRecord {
 
     interface UserDao extends SqlObject {
 
-        @SqlQuery("SELECT * from users order by id")
+        @SqlQuery("""
+                SELECT *
+                FROM users
+                ORDER BY id
+                """)
         List<User> getUsers();
 
-        @SqlQuery("SELECT * from users WHERE id = :id order by id")
+        @SqlQuery("""
+                SELECT *
+                FROM users
+                WHERE id = :id ORDER BY id
+                """)
         Optional<User> getUser(int id);
+
+        @SqlUpdate("""
+                INSERT INTO users (id, name)
+                VALUES (:id, :name)
+                """)
+        void insertUser(@BindMethods User user);
     }
 }
