+++ pom.xml
@@ -99,16 +99,4 @@
             </dependency>
         </dependencies>
     </dependencyManagement>
-
-    <profiles>
-        <profile>
-            <id>java17</id>
-            <activation>
-                <jdk>[17,)</jdk>
-            </activation>
-            <modules>
-                <module>java17</module>
-            </modules>
-        </profile>
-    </profiles>
 </project>
